test: Finish strict-clippy rework of remaining test files
Completes the integration-test migration so `cargo clippy --all-targets
-- -D warnings` and `cargo fmt --check` pass with zero per-item allows.

Test files (migrated to separate-file `#[cfg(test)] mod tests;` form and
fixed for restriction/nursery lints — no `#[allow]`/`#[expect]` added):
- pattern_preprocess_tests, tuple_variant_tests, validation_tests:
  alphabetized fields/variants, removed `println!`, renamed shadowing
  bindings, annotated bare numeric literals, doc punctuation, etc.
- recursive_type_tests: now uses `Vec<Self>` / `HashMap<String, Self>` to
  satisfy `use_self`.

src/ — resolve `Self` in recursive type references:
- Add `FieldDef::resolve_self_references(type_name)` and call it in
  `process_field` (threaded through `collect_struct_fields` and the
  discriminated-enum collector). A field written as `Vec<Self>` is now
  rewritten to the concrete enclosing type name before recursion detection
  and code generation run, so `Self` behaves exactly like the full type
  name and the macro's existing self-reference handling kicks in.

Also normalizes formatting in several previously-committed test files.

Verified: clippy --all-targets clean, fmt --check clean, and the full
`cargo hack test --feature-powerset` passes.
