+++ tests/tuple_variant_tests.rs
@@ -1,703 +1,13 @@
-//! Tests for tuple variant enum support
+//! Tests for tuple variant enum support.
 //!
 //! This module tests the handling of enum tuple variants in TypeScript/Zod/JSON Schema generation:
-//! - Single-element tuples: `Variant(T)` → `{ type: "Variant", value: T }`
-//! - Multi-element tuples: `Variant(T1, T2)` → `{ type: "Variant", value: [T1, T2] }`
-//! - Unit variants in mixed enums: `Variant` → `{ type: "Variant" }`
+//! - Single-element tuples: `Variant(T)` -> `{ type: "Variant", value: T }`
+//! - Multi-element tuples: `Variant(T1, T2)` -> `{ type: "Variant", value: [T1, T2] }`
+//! - Unit variants in mixed enums: `Variant` -> `{ type: "Variant" }`
 //! - Plain enums (all unit variants): string union `"V1" | "V2" | "V3"`
 
-#[cfg(all(test, feature = "typescript", feature = "serde"))]
-mod tests {
-    use serde::{Deserialize, Serialize};
-    use tixschema::model_schema;
-
-    /// Test 1: Single-element tuple variants
-    /// Each variant has exactly one tuple element
-    #[test]
-    fn test_single_tuple_variant_typescript() {
-        #[model_schema()]
-        #[derive(Serialize, Deserialize, Debug, Clone)]
-        pub enum SingleTuple {
-            Text(String),
-            Number(i64),
-            Flag(bool),
-            Decimal(f64),
-        }
-
-        let ts = SingleTuple::ts_definition();
-        println!("Generated TypeScript:\n{ts}");
-
-        // Verify discriminator field
-        assert!(ts.contains("type: \"Text\""), "Missing Text discriminator");
-        assert!(
-            ts.contains("type: \"Number\""),
-            "Missing Number discriminator"
-        );
-        assert!(ts.contains("type: \"Flag\""), "Missing Flag discriminator");
-        assert!(
-            ts.contains("type: \"Decimal\""),
-            "Missing Decimal discriminator"
-        );
-
-        // Verify value field with correct types
-        assert!(ts.contains("value: string"), "Missing string value");
-        assert!(ts.contains("value: number"), "Missing number value");
-        assert!(ts.contains("value: boolean"), "Missing boolean value");
-    }
-
-    /// Test 1b: Single-element tuple variants Zod schema
-    #[cfg(feature = "zod")]
-    #[test]
-    fn test_single_tuple_variant_zod() {
-        #[model_schema()]
-        #[derive(Serialize, Deserialize, Debug, Clone)]
-        pub enum SingleTupleZod {
-            Text(String),
-            Number(i64),
-            Flag(bool),
-        }
-
-        let zod = SingleTupleZod::zod_schema();
-        println!("Generated Zod:\n{zod}");
-
-        // Verify Zod schema structure
-        assert!(
-            zod.contains("z.discriminatedUnion"),
-            "Missing discriminatedUnion"
-        );
-        assert!(zod.contains("z.literal(\"Text\")"), "Missing Text literal");
-        assert!(
-            zod.contains("z.literal(\"Number\")"),
-            "Missing Number literal"
-        );
-        assert!(zod.contains("z.literal(\"Flag\")"), "Missing Flag literal");
-
-        // Verify value types
-        assert!(zod.contains("value: z.string()"), "Missing string value");
-        assert!(
-            zod.contains("value: z.number().int()"),
-            "Missing int value for i64"
-        );
-        assert!(zod.contains("value: z.boolean()"), "Missing boolean value");
-    }
-
-    /// Test 2: Multi-element tuple variants
-    /// Variants with more than one tuple element
-    #[test]
-    fn test_multi_tuple_variant_typescript() {
-        #[model_schema()]
-        #[derive(Serialize, Deserialize, Debug, Clone)]
-        pub enum MultiTuple {
-            Pair(String, i64),
-            Triple(String, i64, bool),
-            Quad(String, i64, bool, f64),
-        }
-
-        let ts = MultiTuple::ts_definition();
-        println!("Generated TypeScript:\n{ts}");
-
-        // Verify tuple syntax
-        assert!(
-            ts.contains("value: [string, number]"),
-            "Missing Pair tuple type"
-        );
-        assert!(
-            ts.contains("value: [string, number, boolean]"),
-            "Missing Triple tuple type"
-        );
-        assert!(
-            ts.contains("value: [string, number, boolean, number]"),
-            "Missing Quad tuple type"
-        );
-    }
-
-    /// Test 2b: Multi-element tuple variants Zod schema
-    #[cfg(feature = "zod")]
-    #[test]
-    fn test_multi_tuple_variant_zod() {
-        #[model_schema()]
-        #[derive(Serialize, Deserialize, Debug, Clone)]
-        pub enum MultiTupleZod {
-            Pair(String, i64),
-            Triple(String, i64, bool),
-        }
-
-        let zod = MultiTupleZod::zod_schema();
-        println!("Generated Zod:\n{zod}");
-
-        // Verify z.tuple() is used
-        assert!(zod.contains("z.tuple("), "Missing z.tuple");
-        assert!(
-            zod.contains("z.tuple([z.string(), z.number().int()])"),
-            "Missing Pair tuple"
-        );
-        assert!(
-            zod.contains("z.tuple([z.string(), z.number().int(), z.boolean()])"),
-            "Missing Triple tuple"
-        );
-    }
-
-    /// Test 3: Plain enum (all unit variants) → string union
-    /// Should NOT generate discriminated union, just string union
-    #[test]
-    fn test_plain_enum_string_union() {
-        #[model_schema()]
-        #[derive(Serialize, Deserialize, Debug, Clone)]
-        pub enum DataType {
-            Alphanumeric,
-            Image,
-            Decimal,
-            Integer,
-            Boolean,
-        }
-
-        let ts = DataType::ts_definition();
-        println!("Generated TypeScript:\n{ts}");
-
-        // Should be a string union, not discriminated union
-        assert!(ts.contains("\"Alphanumeric\""), "Missing Alphanumeric");
-        assert!(ts.contains("\"Image\""), "Missing Image");
-        assert!(ts.contains("\"Decimal\""), "Missing Decimal");
-        assert!(ts.contains("\"Integer\""), "Missing Integer");
-        assert!(ts.contains("\"Boolean\""), "Missing Boolean");
-
-        // Should NOT have type/value fields (it's a plain string union)
-        // The format should be: type DataType = "Alphanumeric" | "Image" | ...
-        assert!(
-            !ts.contains("type:") || ts.contains("export type"),
-            "Should not have type discriminator field"
-        );
-    }
-
-    /// Test 3b: Plain enum Zod schema uses z.enum
-    #[cfg(feature = "zod")]
-    #[test]
-    fn test_plain_enum_zod() {
-        #[model_schema()]
-        #[derive(Serialize, Deserialize, Debug, Clone)]
-        pub enum PlainEnumZod {
-            Active,
-            Inactive,
-            Pending,
-        }
-
-        let zod = PlainEnumZod::zod_schema();
-        println!("Generated Zod:\n{zod}");
-
-        // Should use z.enum, not z.discriminatedUnion
-        assert!(zod.contains("z.enum("), "Should use z.enum for plain enums");
-        assert!(
-            !zod.contains("z.discriminatedUnion"),
-            "Should not use discriminatedUnion for plain enums"
-        );
-    }
-
-    /// Test 4: Mixed variants (comprehensive)
-    /// Mix of unit, tuple-single, tuple-multi, and named struct variants
-    #[test]
-    fn test_mixed_variants_typescript() {
-        #[model_schema()]
-        #[derive(Serialize, Deserialize, Debug, Clone)]
-        pub enum Mixed {
-            // Unit variant
-            Empty,
-            // Single tuple
-            Text(String),
-            // Multi tuple
-            Pair(String, i64),
-            // Named struct
-            Named { field_a: String, field_b: bool },
-        }
-
-        let ts = Mixed::ts_definition();
-        println!("Generated TypeScript:\n{ts}");
-
-        // Unit variant should only have discriminator
-        assert!(
-            ts.contains("type: \"Empty\""),
-            "Missing Empty discriminator"
-        );
-
-        // Single tuple should have value field
-        assert!(ts.contains("type: \"Text\""), "Missing Text discriminator");
-        assert!(ts.contains("value: string"), "Missing Text value");
-
-        // Multi tuple should have tuple value
-        assert!(ts.contains("type: \"Pair\""), "Missing Pair discriminator");
-        assert!(
-            ts.contains("value: [string, number]"),
-            "Missing Pair tuple value"
-        );
-
-        // Named struct should have individual fields
-        assert!(
-            ts.contains("type: \"Named\""),
-            "Missing Named discriminator"
-        );
-        assert!(ts.contains("field_a: string"), "Missing field_a");
-        assert!(ts.contains("field_b: boolean"), "Missing field_b");
-    }
-
-    /// Test 4b: Mixed variants Zod schema
-    #[cfg(feature = "zod")]
-    #[test]
-    fn test_mixed_variants_zod() {
-        #[model_schema()]
-        #[derive(Serialize, Deserialize, Debug, Clone)]
-        pub enum MixedZod {
-            Empty,
-            Text(String),
-            Pair(String, i64),
-            Named { field_a: String, field_b: bool },
-        }
-
-        let zod = MixedZod::zod_schema();
-        println!("Generated Zod:\n{zod}");
-
-        // Should use discriminatedUnion since not all variants are unit
-        assert!(
-            zod.contains("z.discriminatedUnion"),
-            "Should use discriminatedUnion"
-        );
-
-        // Unit variant
-        assert!(
-            zod.contains("z.literal(\"Empty\")"),
-            "Missing Empty literal"
-        );
-
-        // Single tuple
-        assert!(zod.contains("z.literal(\"Text\")"), "Missing Text literal");
-        assert!(zod.contains("value: z.string()"), "Missing Text value");
-
-        // Multi tuple
-        assert!(zod.contains("z.literal(\"Pair\")"), "Missing Pair literal");
-        assert!(zod.contains("z.tuple("), "Missing tuple");
-
-        // Named struct
-        assert!(
-            zod.contains("z.literal(\"Named\")"),
-            "Missing Named literal"
-        );
-        assert!(zod.contains("field_a: z.string()"), "Missing field_a");
-        assert!(zod.contains("field_b: z.boolean()"), "Missing field_b");
-    }
-
-    /// Test 5: JSON Schema generation for tuple variants
-    #[cfg(feature = "jsonschema")]
-    #[test]
-    fn test_tuple_json_schema() {
-        #[model_schema()]
-        #[derive(Serialize, Deserialize, Debug, Clone)]
-        pub enum TupleSchema {
-            Single(String),
-            Double(String, i64),
-        }
-
-        let schema = TupleSchema::json_schema();
-        let schema_str = serde_json::to_string_pretty(&schema).unwrap();
-        println!("Generated JSON Schema:\n{schema_str}");
-
-        // Should have oneOf for discriminated union
-        assert!(schema_str.contains("\"oneOf\""), "Missing oneOf");
-
-        // Should have value property
-        assert!(schema_str.contains("\"value\""), "Missing value property");
-
-        // Multi-tuple should use prefixItems
-        assert!(
-            schema_str.contains("prefixItems"),
-            "Multi-tuple should use prefixItems"
-        );
-    }
-
-    /// Test 6: Custom content field name via serde
-    #[test]
-    fn test_custom_content_field() {
-        #[model_schema()]
-        #[derive(Serialize, Deserialize, Debug, Clone)]
-        #[serde(tag = "kind", content = "data")]
-        pub enum CustomContent {
-            Text(String),
-            Number(i64),
-        }
-
-        let ts = CustomContent::ts_definition();
-        println!("Generated TypeScript:\n{ts}");
-
-        // Should use "kind" instead of "type"
-        assert!(ts.contains("kind: \"Text\""), "Should use 'kind' as tag");
-        assert!(ts.contains("kind: \"Number\""), "Should use 'kind' as tag");
-
-        // Should use "data" instead of "value"
-        assert!(
-            ts.contains("data: string"),
-            "Should use 'data' as content field"
-        );
-        assert!(
-            ts.contains("data: number"),
-            "Should use 'data' as content field"
-        );
-    }
-
-    /// Test 6b: Custom content field Zod schema
-    #[cfg(feature = "zod")]
-    #[test]
-    fn test_custom_content_field_zod() {
-        #[model_schema()]
-        #[derive(Serialize, Deserialize, Debug, Clone)]
-        #[serde(tag = "kind", content = "data")]
-        pub enum CustomContentZod {
-            Text(String),
-            Number(i64),
-        }
-
-        let zod = CustomContentZod::zod_schema();
-        println!("Generated Zod:\n{zod}");
-
-        // Should use "kind" in discriminatedUnion
-        assert!(
-            zod.contains("z.discriminatedUnion(\"kind\""),
-            "Should use 'kind' as discriminator"
-        );
-
-        // Should use "data" as field name
-        assert!(
-            zod.contains("data: z.string()"),
-            "Should use 'data' as content field"
-        );
-        assert!(
-            zod.contains("data: z.number()"),
-            "Should use 'data' as content field"
-        );
-    }
-
-    /// Test 7: Tuple variant with Vec (array type in tuple)
-    #[test]
-    fn test_tuple_with_vec() {
-        #[model_schema()]
-        #[derive(Serialize, Deserialize, Debug, Clone)]
-        pub enum TupleWithVec {
-            // This was the original problem case: Image(String, Vec<u8>)
-            Image(String, Vec<u8>),
-            Data(Vec<String>),
-        }
-
-        let ts = TupleWithVec::ts_definition();
-        println!("Generated TypeScript:\n{ts}");
-
-        // Image should have tuple with string and array
-        assert!(
-            ts.contains("value: [string, Array<number>]"),
-            "Image should have [string, Array<number>]"
-        );
-
-        // Data should have array value (single tuple)
-        assert!(
-            ts.contains("value: Array<string>"),
-            "Data should have Array<string>"
-        );
-    }
-
-    /// Test 7b: Tuple with Vec Zod schema
-    #[cfg(feature = "zod")]
-    #[test]
-    fn test_tuple_with_vec_zod() {
-        #[model_schema()]
-        #[derive(Serialize, Deserialize, Debug, Clone)]
-        pub enum TupleWithVecZod {
-            Image(String, Vec<u8>),
-            Data(Vec<String>),
-        }
-
-        let zod = TupleWithVecZod::zod_schema();
-        println!("Generated Zod:\n{zod}");
-
-        // Image tuple
-        assert!(
-            zod.contains("z.tuple([z.string(), z.array(z.number().int())])"),
-            "Image should have z.tuple"
-        );
-
-        // Data single tuple
-        assert!(
-            zod.contains("value: z.array(z.string())"),
-            "Data should have z.array"
-        );
-    }
-
-    /// Test 8: Regression - discriminated union with named struct still works
-    #[test]
-    fn test_named_struct_regression() {
-        #[model_schema()]
-        #[derive(Serialize, Deserialize, Debug, Clone)]
-        pub enum PaymentMethod {
-            CreditCard {
-                card_number: String,
-                expiry: String,
-            },
-            BankTransfer {
-                account_number: String,
-                routing_number: String,
-            },
-        }
-
-        let ts = PaymentMethod::ts_definition();
-        println!("Generated TypeScript:\n{ts}");
-
-        // Should still work as before
-        assert!(
-            ts.contains("type: \"CreditCard\""),
-            "Missing CreditCard discriminator"
-        );
-        assert!(ts.contains("card_number: string"), "Missing card_number");
-        assert!(ts.contains("expiry: string"), "Missing expiry");
-
-        assert!(
-            ts.contains("type: \"BankTransfer\""),
-            "Missing BankTransfer discriminator"
-        );
-        assert!(
-            ts.contains("account_number: string"),
-            "Missing account_number"
-        );
-        assert!(
-            ts.contains("routing_number: string"),
-            "Missing routing_number"
-        );
-    }
-
-    /// Test 9: Optional types in tuple variants
-    #[test]
-    fn test_optional_in_tuple() {
-        #[model_schema()]
-        #[derive(Serialize, Deserialize, Debug, Clone)]
-        pub enum OptionalTuple {
-            Maybe(Option<String>),
-            MaybePair(String, Option<i64>),
-        }
-
-        let ts = OptionalTuple::ts_definition();
-        println!("Generated TypeScript:\n{ts}");
-
-        // Single optional tuple
-        assert!(
-            ts.contains("value: string | undefined"),
-            "Maybe should have optional string"
-        );
-
-        // Multi tuple with optional element
-        assert!(
-            ts.contains("[string, number | undefined]"),
-            "MaybePair should have tuple with optional element"
-        );
-    }
-
-    /// Test 10: Tuple variant with nested custom type
-    #[test]
-    fn test_tuple_with_custom_type() {
-        #[model_schema()]
-        #[derive(Serialize, Deserialize, Debug, Clone)]
-        pub struct Inner {
-            pub field: String,
-        }
-
-        #[model_schema()]
-        #[derive(Serialize, Deserialize, Debug, Clone)]
-        pub enum Outer {
-            Wrapped(Inner),
-            Paired(Inner, i64),
-        }
-
-        let ts = Outer::ts_definition();
-        println!("Generated TypeScript:\n{ts}");
-
-        // Should reference the inner type
-        assert!(
-            ts.contains("value: Inner"),
-            "Wrapped should reference Inner type"
-        );
-        assert!(
-            ts.contains("value: [Inner, number]"),
-            "Paired should have tuple with Inner"
-        );
-    }
-
-    /// Test 11: Serde serialization compatibility
-    /// Verify that the generated schema matches actual serde serialization
-    #[test]
-    fn test_serde_serialization_compatibility() {
-        #[model_schema()]
-        #[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
-        #[serde(tag = "type", content = "value")]
-        pub enum SerdeCompat {
-            Text(String),
-            Number(i64),
-            Pair(String, i64),
-        }
-
-        // Test serialization produces expected format
-        let text = SerdeCompat::Text("hello".to_owned());
-        let json = serde_json::to_value(&text).unwrap();
-        assert_eq!(json["type"], "Text");
-        assert_eq!(json["value"], "hello");
-
-        let number = SerdeCompat::Number(42);
-        let json = serde_json::to_value(&number).unwrap();
-        assert_eq!(json["type"], "Number");
-        assert_eq!(json["value"], 42);
-
-        let pair = SerdeCompat::Pair("hello".to_owned(), 42);
-        let json = serde_json::to_value(&pair).unwrap();
-        assert_eq!(json["type"], "Pair");
-        assert!(json["value"].is_array());
-        assert_eq!(json["value"][0], "hello");
-        assert_eq!(json["value"][1], 42);
-    }
-
-    /// Test 12: Complex `FixedValue` enum from original issue
-    /// This is the exact enum from the user's original problem
-    #[test]
-    fn test_fixed_value_original_issue() {
-        #[model_schema()]
-        #[derive(Serialize, Deserialize, Debug, Clone)]
-        pub enum FixedValue {
-            Alphanumeric(String),
-            Image(String, Vec<u8>),
-            Decimal(f64),
-            Integer(i64),
-            Boolean(bool),
-        }
-
-        let ts = FixedValue::ts_definition();
-        println!("Generated TypeScript:\n{ts}");
-
-        // Verify NO empty field names (the original bug)
-        // Empty field names would look like "  : string;" (space before colon, no field name)
-        // Valid output has field names like "value: string;"
-        assert!(
-            !ts.contains("\n  : "),
-            "Should not have empty field name (found line starting with colon)"
-        );
-        // Also verify we don't have the old pattern of empty property names in the schema
-        assert!(
-            !ts.contains("\"\":"),
-            "JSON Schema should not have empty property names"
-        );
-
-        // Verify proper structure
-        assert!(
-            ts.contains("type: \"Alphanumeric\""),
-            "Missing Alphanumeric"
-        );
-        assert!(
-            ts.contains("value: string"),
-            "Alphanumeric should have value"
-        );
-
-        assert!(ts.contains("type: \"Image\""), "Missing Image");
-        assert!(
-            ts.contains("[string, Array<number>]"),
-            "Image should have tuple"
-        );
-
-        assert!(ts.contains("type: \"Decimal\""), "Missing Decimal");
-        assert!(ts.contains("type: \"Integer\""), "Missing Integer");
-        assert!(ts.contains("type: \"Boolean\""), "Missing Boolean");
-    }
-
-    /// Test 13: `FixedValueExt` with all variant types
-    #[test]
-    fn test_fixed_value_ext_comprehensive() {
-        #[model_schema()]
-        #[derive(Serialize, Deserialize, Debug, Clone)]
-        pub enum FixedValueExt {
-            Alphanumeric(String),
-            Image(String, Vec<u8>),
-            Decimal(f64),
-            Integer(i64),
-            Boolean(bool),
-            Tuple(i64, bool),
-            SingleValue,
-            Complex { a: String, b: bool },
-        }
-
-        let ts = FixedValueExt::ts_definition();
-        println!("Generated TypeScript:\n{ts}");
-
-        // Single tuple variants
-        assert!(
-            ts.contains("type: \"Alphanumeric\""),
-            "Missing Alphanumeric"
-        );
-        assert!(
-            ts.contains("value: string") || ts.contains("value: number"),
-            "Single tuple should have value"
-        );
-
-        // Multi-element tuple
-        assert!(ts.contains("type: \"Image\""), "Missing Image");
-        assert!(ts.contains("type: \"Tuple\""), "Missing Tuple");
-        assert!(
-            ts.contains("[number, boolean]"),
-            "Tuple should have [number, boolean]"
-        );
-
-        // Unit variant
-        assert!(ts.contains("type: \"SingleValue\""), "Missing SingleValue");
-
-        // Named struct variant
-        assert!(ts.contains("type: \"Complex\""), "Missing Complex");
-        assert!(ts.contains("a: string"), "Missing field a");
-        assert!(ts.contains("b: boolean"), "Missing field b");
-    }
-
-    /// Test 14: Empty tuple variant (edge case)
-    /// An empty tuple `Foo()` should be treated like a unit variant
-    #[test]
-    fn test_empty_tuple_variant() {
-        #[model_schema()]
-        #[derive(Serialize, Deserialize, Debug, Clone)]
-        pub enum EmptyTuple {
-            Normal(String),
-            // Note: Rust doesn't allow `Empty()` syntax directly, but we test the logic anyway
-            // by having a unit variant alongside tuple variants
-            Unit,
-        }
-
-        let ts = EmptyTuple::ts_definition();
-        println!("Generated TypeScript:\n{ts}");
-
-        // Both should be in discriminated union
-        assert!(ts.contains("type: \"Normal\""), "Missing Normal");
-        assert!(ts.contains("type: \"Unit\""), "Missing Unit");
-
-        // Unit should not have value field
-        // The Unit variant's object should just have `type: "Unit"` without a value field
-        // This is verified by ensuring the total schema structure is correct
-    }
-
-    /// Test 15: `JSDoc` comments in generated TypeScript
-    #[test]
-    fn test_jsdoc_comments() {
-        #[model_schema()]
-        #[derive(Serialize, Deserialize, Debug, Clone)]
-        pub enum JsDoc {
-            Single(String),
-            Multi(String, i64),
-        }
-
-        let ts = JsDoc::ts_definition();
-        println!("Generated TypeScript:\n{ts}");
-
-        // Should have JSDoc comments
-        assert!(ts.contains("/**"), "Should have JSDoc comments");
-        assert!(ts.contains("*/"), "Should have JSDoc end");
-
-        // Multi-tuple should have tuple description
-        assert!(
-            ts.contains("Tuple:") || ts.contains("Tuple value"),
-            "Multi-tuple should have tuple documentation"
-        );
-    }
-}
+#[cfg(test)]
+#[cfg(feature = "typescript")]
+#[cfg(feature = "serde")]
+#[path = "tuple_variant_tests/tests.rs"]
+mod tests;
